feat(ST41): Build UI for vocabulary boost screen (Prototype 04)
